Create antiHoneyPot.js
You will need to replace YOUR_PROJECT_ID with your Infura project ID, and contractAddress with the address of the smart contract you're interested in.

This script uses the Web3.js library to interact with the Ethereum blockchain, and the Infura API as a provider to connect to the Ethereum network. The script calls the balanceOf function of the smart contract to get the balance of the specified address and print the result on the console.
This script is still in an implementation stage add more functionality and error handling capabilities are yet to be included.